requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
gspread==5.12.0
google-auth==2.27.0
python-dotenv==1.0.0
requests==2.31.0 